refactor(components/atom/input): move some files - rename diretories
