+++ components/atom/input/src/Input/index.js
@@ -1,6 +1,6 @@
-import withAddons from './Features/Addon'
-import withIcons from './Features/Icon'
-import withButton from './Features/WithButton'
+import withAddons from './HOC/withAddons'
+import withIcons from './HOC/withIcons'
+import withButton from './HOC/withButton'
 import Input, {inputStates, inputSizes} from './Component'
 
 export default withButton(withAddons(withIcons(Input)))
